Renamed app to a more generic name, since the idea is that the data model it describes is flexible enough to be extendable to other majors in the UPF.
The parsing will be specific for each major. There will be a script that will take care of the ESUP and, in the future, maybe I will add others.

The models from the app have received a major overhaul. I think that I won't need more fields nor models, but I still need to describe what kind of queries can be performed on the data.

Other changes are simply to change the module name where needed.