Created subjectparser command.
Parses the list of subjects of the ESUP (right now, only ESUP). Has a signal to add duplicates onto a table, so every time a subject is deleted (there may be 1 subject with multiple names due to mispellings, etc.) it is added onto the duplicates database, so if you run the script again, the duplicates won't be inserted.