New alias and lesson parser. Refactored to take new timetable.json as input for storing data.
Lessonparser should be inspected carefully, since it has non obvious bugs that make the script crash. The parser also has to be revisited, and refactored so it raises more exceptions, and clearer ones too. I should also include a sort of system to report those exceptions via email. This has to happen. since the parser may not be able to figure out some of the additions to the calendar.