Major change.
This deletes a lot of models in the database to deal with Lesson insertion and deletion (keeping track of what was done, to avoid destructive changes)

I have switched to a simpler model, where there are no Lesson kinds (theory, seminar, etc.) or subgroups. Right now, this only divides lessons by subject and time, and puts everything else that appears in the original source. It’s way more stupid, but much more reliable, and makes my life easier and maintenance minimum (i hope).

It also has the additional benefit that I have been able to delete a ton of code, which is nice.