Changed the method used to query lessons to be deleted.
Now, a uuid field is added to the Lesson model. This uuid is the hash() of a tuple containing the lesson object (as parsed by timetable.parser), group and academic year.

The nice thing about this method is that, even If I change something manually, the entry is still linked to that parsed Lesson object.

Things that ain’t so hot: It may happen that 2 different timetables, referring to the same entry, by being slightly different will get duplicated, but I think that may also happen in the previous system, so no (many) worries.

Anyway, since we have the uuid field, deleting lessons is as simple as performing a get for that uuid, with very little change to screw things.