Hour parser fix. Should be more resilient to bad inputs now (still not perfect though).
Trying to prevent inputs like

54.S02/S05/S06

To be parsed as an hour. With the fix right now, this works since 54 is > 24, so it can't be an hour, but if the building was a valid hour, it would still pass. In this case, since all ESUP building numbers are higher than that we are safe (but it's a shitty fix).