This is finally a good fix for the asyncio scraper. Instead of having a pool of sessions which are reused, what it does now is follow a simpler model. Each time we need to make a request, we create a session, init that session, make the requests, then close the session. Drawbacks: You have to open a new session every time, which is a new connection to the server, new https handshake, new GET request. It's more expensive, time wise. Pros: Works perfectly. Connections properly closed, less code, more clear. Concurrency level simply controlled by a semaphore, no problem trying to use same session concurrently, etc. etc.
Overall, for now, even if it may not be the fastest solution, it still should be quite a bit more fast. I'll test afterwards and run some numbers to quantify the game.

Anyhow, this is an actual scraper using asnycio that works, for something less trivial than just running some independent HTTP requests